backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pymupdf
scikit-learn
spacy
python-docx